refactor: factor out feedback gen from matcher
